refactor compression sdk
