+++ src/sdk/pynni/nni/compression/tensorflow/__init__.py
@@ -1,0 +1,3 @@
+from .compressor import LayerInfo, Compressor, Pruner, Quantizer
+from .builtin_pruners import *
+from .builtin_quantizers import *
