+++ src/sdk/pynni/nni/compression/tensorflow/default_layers.py
@@ -1,0 +1,8 @@
+op_weight_index = {
+    'Conv2D': None,
+    'Conv3D': None,
+    'DepthwiseConv2dNative': None,
+
+    'Mul': None,
+    'MatMul': None,
+}
