--- src/sdk/pynni/nni/compressors/tf_compressor/__init__.py
@@ -1,3 +1,0 @@
-from .pruner import LevelPruner, AGPruner, SensitivityPruner
-from .quantizer import NaiveQuantizer, QATquantizer, DoReFaQuantizer
-from ._nnimc_tf import TfPruner, TfQuantizer
