--- src/sdk/pynni/nni/compressors/torch_compressor/__init__.py
@@ -1,3 +1,0 @@
-from .pruner import LevelPruner, AGPruner, SensitivityPruner
-from .quantizer import NaiveQuantizer, DoReFaQuantizer, QATquantizer
-from ._nnimc_torch import TorchPruner, TorchQuantizer
